Save discount data in Cart during checkout
