+++ saleor/cart/checkout/forms.py
@@ -1,10 +1,15 @@
+from datetime import date
+
 from django import forms
+from django.utils.encoding import smart_text
 from django.utils.safestring import mark_safe
 from django.utils.translation import pgettext_lazy
 
+from ...cart.checkout.utils import get_voucher_discount_for_checkout
 from ...cart.models import Cart
 from ...core.utils import format_money
 from ...core.utils.taxes import display_gross_prices
+from ...discount.models import NotApplicable, Voucher
 from ...shipping.models import ShippingMethodCountry
 from ...shipping.utils import get_taxed_shipping_price
 
@@ -137,3 +142,56 @@ class CartNoteForm(forms.ModelForm):
     class Meta:
         model = Cart
         fields = ['note']
+
+
+class VoucherField(forms.ModelChoiceField):
+
+    default_error_messages = {
+        'invalid_choice': pgettext_lazy(
+            'Voucher form error', 'Discount code incorrect or expired'),
+    }
+
+
+class CheckoutDiscountForm(forms.ModelForm):
+
+    voucher = VoucherField(
+        queryset=Voucher.objects.none(),
+        to_field_name='code',
+        help_text=pgettext_lazy(
+            'Checkout discount form label for voucher field',
+            'Gift card or discount code'),
+        widget=forms.TextInput)
+
+    class Meta:
+        model = Cart
+        fields = ['voucher']
+
+    def __init__(self, *args, **kwargs):
+        self.checkout = kwargs.pop('checkout')
+        initial = kwargs.get('initial', {})
+        if 'voucher' not in initial:
+            initial['voucher'] = self.instance.voucher_code
+        kwargs['initial'] = initial
+        super().__init__(*args, **kwargs)
+        self.fields['voucher'].queryset = Voucher.objects.active(
+            date=date.today())
+
+    def clean(self):
+        cleaned_data = super().clean()
+        if 'voucher' in cleaned_data:
+            voucher = cleaned_data['voucher']
+            try:
+                discount = get_voucher_discount_for_checkout(
+                    voucher, self.checkout)
+                cleaned_data['discount'] = discount
+                cleaned_data['discount_name'] = voucher.name
+            except NotApplicable as e:
+                self.add_error('voucher', smart_text(e))
+        return cleaned_data
+
+    def save(self, commit=True):
+        voucher = self.cleaned_data['voucher']
+        self.instance.voucher_code = voucher.code
+        self.instance.discount_amount = self.cleaned_data['discount']
+        self.instance.discount_name = self.cleaned_data['discount_name']
+        return super().save(commit)
